TP1/requirements.txt
pillow==10.2.0
numpy==2.0.1
scipy==1.14.0